fix stall of poll fd
ZMQ documentation states that the fd handle will be edge-notified of any event on the socket by signaling a readable event on the fd.
this fixed two errors in this code based on that documentation:
1) it is always signaled as a read, even for a recv operation
2) since the fd is not notified if there is an event pending immediately after a zmq_send or zmq_recv, call notify manually at that point to wake up any other current watchers

other bug fixes:
null the data field during close earlier. if any errors happen, the socket should still be considered dead from the julia side
make send nonblocking (like recv), and change the flags parameter to a boolean (for SNDMORE)
call zmq_errno instead of system errno, in case zmq is linked against a different libc than julia